requirements.txt
Django==5.1.2
djangorestframework==3.15.2
requests==2.32.3
pydantic
python-dotenv
pydantic[dotenv]
drf-yasg
psycopg2
psycopg2-binary
psycopg[binary]
django-filter